feat: Add 5 max tries for scraping current page
When the underlying container element for the job cards does not load (div.job_seen_beacon) after 5 seconds, the Selenium driver will refresh and try to reload the page. The driver will do this for a max of 5 attempts before short-circuiting. This feature improves consistency and accuracy of the resulting dataset.