feat: Add logic for 'applied' header
Added the 'applied' header to the resulting dataset to help users keep track of jobs they have already applied to. If a previously search result appears in new webscrape attempts, the hash_id will be used to identify whether the job has been applied to already by the user. This feature introduces tracking via 'applied' and enables the user to convert the dataset into a working spreadsheet if desired.