feat: Scraping loop optimization
The web scraper will stop scraping the job results if the resulting hash_ids from the current page are exactly the same as the previous page. The matching hash_ids indicate that the parser has reached the last page of search results (using Indeed). This feature decreases unnecessary web traffic and processing time.