feat: Reconfigured default config
This commit includes the 'applied' attribute in the resulting dataset and configures the web parser to scrape all results by default (instead of a specified number of pages). Additional keywords were added for preference (these are encouraged to be modified by all users to adjust their scraped results).

The resulting data set is a result of running the implemented scraper with the configurations adjusted in the config.json.