feat: Implement GUI foundation
This commit introduces a GUI for the user to modify the configuration settings of the scraper. The 'Indeed' setting fields and drop-downs are available for the user to interact with when running the MainFrame.py file. Prepare for integrating a frame for excluded keywords and another frame for modifying csv settings.